pyproject.toml
[tool.black]
line-length = 100

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = ["app/migrations/env.py"]

[tool.poetry]
name = "BroCast"
version = "1.5.0"
description = "BroCast! The chat app where you can chat with your bros!"
authors = ["Zwaar Developers"]
license = "GPL-3.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
fastapi = "0.115.6"
uvicorn = "0.34.0"
python-dotenv = "1.0.1"
psycopg2-binary = "2.9.10"
sqlmodel = "0.0.22"
alembic = "1.14.0"
asyncpg = "0.30.0"
passlib = "1.7.4"
authlib = "1.4.0"
oauthlib = "3.2.2"
python-socketio = "5.12.1"
websockets = "14.1"
asyncio-redis = "0.16.0"
redis = "5.2.1"
pillow = "11.1.0"
fastapi-pagination = "0.12.34"
fastapi-mail = "1.4.2"
celery = "5.4.0"
APScheduler = "3.11.0"
pydantic_settings = "2.7.1"
requests = "2.32.3"
PyJWT = "2.10.1"
python-multipart = "0.0.20"
pytz = "2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
